Stochastic Gradient Hamiltonian Monte Carlo

Simulates from the posterior defined by the functions logLik and logPrior using stochastic gradient Hamiltonian Monte Carlo. The function uses TensorFlow, so needs Tensorflow for python installed. Currently we use the approximation \(\hat \beta = 0\), as used in the simulations by the original reference. This will be changed in future implementations.

sghmc(logLik, dataset, params, stepsize, logPrior = NULL,
  minibatchSize = 0.01, alpha = 0.01, L = 5L, nIters = 10^4L,
  verbose = TRUE)

Arguments

logLik
function which takes parameters and dataset (list of TensorFlow variables and placeholders respectively) as input. It should return a TensorFlow expression which defines the log likelihood of the model.

dataset
list of numeric R arrays which defines the datasets for the problem. The names in the list should correspond to those referred to in the logLik and logPrior functions

params
list of numeric R arrays which define the starting point of each parameter. The names in the list should correspond to those referred to in the logLik and logPrior functions

stepsize
list of numeric values corresponding to the SGHMC stepsizes for each parameter (\(\eta\) in the original paper). The names in the list should correspond to those in params. Alternatively specify a single numeric value to use that stepsize for all parameters.

logPrior
optional. Default uninformative improper prior. Function which takes parameters (list of TensorFlow variables) as input. The function should return a TensorFlow tensor which defines the log prior of the model.

minibatchSize
optional. Default 0.01. Numeric or integer value that specifies amount of dataset to use at each iteration either as proportion of dataset size (if between 0 and 1) or actual magnitude (if an integer).

alpha
optional. Default 0.01. List of numeric values corresponding to the SGHMC momentum tuning constants (\(\alpha\) in the original paper). One value should be given for each parameter in params, the names should correspond to those in params. Alternatively specify a single float to specify that value for all parameters.

L
optional. Default 5L. Integer specifying the trajectory parameter of the simulation, as defined in the main reference.

nIters
optional. Default 10^4L. Integer specifying number of iterations to perform.

verbose
optional. Default TRUE. Boolean specifying whether to print algorithm progress

Value

Returns list of arrays for each parameter containing the MCMC chain. Dimension of the form (nIters,paramDim1,paramDim2,...)

References

Chen, T., Fox, E. B., and Guestrin, C. (2014). Stochastic gradient Hamiltonian Monte Carlo. In ICML (pp. 1683-1691).