Stochastic Gradient Hamiltonian Monte Carlo with Control Variates

Simulates from the posterior defined by the functions logLik & logPrior using stochastic gradient Hamiltonian Monte Carlo with an improved gradient estimate that is calculated using control variates. Currently we use the approximation \(\hat \beta = 0\), as used in the simulations by the original reference. This will be changed in future implementations.

sghmccv(logLik, logPrior, dataset, params, eta, alpha, L, optStepsize, n,
  nIters = 10^4, nItersOpt = 10^4, verbose = TRUE)

Arguments

logLik
function which takes parameters and dataset (list of tensorflow variables and placeholders respectively) as input. It should return a tensorflow expression which defines the log likelihood of the model.

logPrior
function which takes parameters and dataset (list of tensorflow variables and placeholders respectively) as input. The function should return a tensorflow tensor which defines the log prior of the model.

dataset
list of R arrays which defines the datasets for the problem. The names in the list should correspond to those referred to in the logLik and logPrior functions

params
list of R arrays which define the starting point of each parameter. The names in the list should correspond to those referred to in the logLik and logPrior functions

eta
list of numeric values corresponding to the SGHMC eta stepsize terms, as defined in the paper given in the references. One value should be given for each parameter in params, the names should correspond to those in params.

alpha
list of numeric values corresponding to the SGHMC alpha momentum terms, as defined in the paper given in the references. One value should be given for each parameter in params, the names should correspond to those in params.

L
integer specifying the trajectory of the simulation, as defined in the main reference.

optStepsize
numeric, stepsize for optimization step to find MAP estimates of parameters.

n
minibatch size, assumed integer.

nIters
integer, number of iterations of SGLD to perform, optional, default 10^4.

nItersOpt
integer, number of iterations of initial optimization to perform, optional, default 10^4.

verbose
Boolean, whether to print algorithm progress or not, default TRUE.

Value

List of arrays for each parameter containing the MCMC chain. Dimension of the form (nIters,paramDim1,paramDim2,...)

References

Baker, J., Fearnhead, P., Fox, E. B., & Nemeth, C. (2017). Control variates for stochastic gradient MCMC. ArXiv preprint arXiv:1706.05439.

Chen, T., Fox, E. B., & Guestrin, C. (2014). Stochastic gradient Hamiltonian Monte Carlo. In ICML (pp. 1683-1691).