Convert R Object into TensorFlow Object

In most cases R objects will be automatically converted to TensorFlow objects in a natural way. But we can sometimes run into issues when these are converted in an unexpected way. For example, the type tf$float64 is not as well supported as tf$float32, so we write everything assuming we are working with tf$float32 types. This can sometimes be a problem if an R object in your LogLik or LogPrior functions is converted to tf$float64 type or tf$int64 object. While this is not normally the case it can sometimes happen, and the way to get around it is to explicitly convert that object to a tensorflow inside the logLik or logPrior functions and declare its type. Which is done using the tf$constant function declared below

Arguments

object
R object such as a numeric, matrix or array object.

dtype
The TensorFlow type of the object, ensure this is 32-bit so tf$float32 or tf$int32.

Value

Returns a tensorflow object with the declared dtype which can be used as input to other TensorFlow functions.